Move burrito endpoints to burrito2 dir
