+++ src/endpoints/burrito2/get_ingredient_prettified.rs
@@ -1,0 +1,74 @@
+use std::path::{Components, PathBuf};
+use rocket::{get, State};
+use rocket::http::{ContentType, Status};
+use rocket::response::status;
+use crate::structs::AppSettings;
+use crate::utils::json_responses::make_bad_json_data_response;
+use crate::utils::paths::{check_path_components, check_path_string_components, os_slash_str};
+
+/// *`GET /ingredient/prettified/<repo_path>?ipath=my_burrito_path`*
+///
+/// Typically mounted as **`/burrito/ingredient/prettified/<repo_path>?ipath=my_burrito_path`**
+///
+/// Returns a text-like resource as a web page.
+#[get("/ingredient/prettified/<repo_path..>?<ipath>")]
+pub async fn get_ingredient_prettified(
+    state: &State<AppSettings>,
+    repo_path: PathBuf,
+    ipath: String,
+) -> status::Custom<(ContentType, String)> {
+    let path_components: Components<'_> = repo_path.components();
+    if check_path_components(&mut path_components.clone())
+        && check_path_string_components(ipath.clone())
+    {
+        let path_to_serve = state.repo_dir.lock().unwrap().clone()
+            + os_slash_str()
+            + &repo_path.display().to_string()
+            + "/ingredients/"
+            + ipath.as_str();
+        let file_string = match std::fs::read_to_string(path_to_serve) {
+            Ok(v) => v,
+            Err(e) => {
+                return status::Custom(
+                    Status::BadRequest,
+                    (
+                        ContentType::JSON,
+                        make_bad_json_data_response(
+                            format!("could not read ingredient content: {}", e).to_string(),
+                        ),
+                    ),
+                )
+            }
+        };
+        status::Custom(
+            Status::Ok,
+            (
+                ContentType::HTML,
+                format!(
+                    r#"
+                <html>
+                <head>
+                <title>Prettified</title>
+                <link rel="stylesheet" href="/webfonts/_webfonts.css">
+                </head>
+                <body>
+                <pre>
+                {}
+                </pre>
+                </body>
+                </html>
+                "#,
+                    file_string
+                ),
+            ),
+        )
+    } else {
+        status::Custom(
+            Status::BadRequest,
+            (
+                ContentType::JSON,
+                make_bad_json_data_response("bad repo path".to_string()),
+            ),
+        )
+    }
+}
