+++ src/endpoints/burrito2/post_raw_ingredient.rs
@@ -1,0 +1,71 @@
+use std::path::{Components, PathBuf};
+use rocket::{post, State};
+use rocket::http::{ContentType, Status};
+use rocket::response::status;
+use rocket::serde::json::Json;
+use serde_json::Value;
+use crate::structs::AppSettings;
+use crate::utils::json_responses::{make_bad_json_data_response, make_good_json_data_response};
+use crate::utils::paths::{check_path_components, check_path_string_components, os_slash_str};
+
+/// *`POST /ingredient/raw/<repo_path>?ipath=my_burrito_path`*
+///
+/// Typically mounted as **`/burrito/ingredient/raw/<repo_path>?ipath=my_burrito_path`**
+///
+/// Writes a document, where the document is provided as JSON with a 'payload' key.
+///
+/// /// The target file must exist, ie this is not the way to add new ingredients to a Burrito
+#[post(
+    "/ingredient/raw/<repo_path..>?<ipath>",
+    format = "json",
+    data = "<json_form>"
+)]
+pub async fn post_raw_ingredient(
+    state: &State<AppSettings>,
+    repo_path: PathBuf,
+    ipath: String,
+    json_form: Json<Value>,
+) -> status::Custom<(ContentType, String)> {
+    let path_components: Components<'_> = repo_path.components();
+    let destination = state.repo_dir.lock().unwrap().clone()
+        + os_slash_str()
+        + &repo_path.display().to_string()
+        + "/ingredients/"
+        + ipath.clone().as_str();
+    if check_path_components(&mut path_components.clone())
+        && check_path_string_components(ipath.clone())
+        && std::fs::metadata(destination.clone()).is_ok()
+    {
+        match std::fs::write(destination, json_form["payload"].as_str().unwrap()) {
+            Ok(_) => status::Custom(
+                Status::Ok,
+                (
+                    ContentType::JSON,
+                    make_good_json_data_response("ok".to_string()),
+                ),
+            ),
+            Err(e) => status::Custom(
+                Status::InternalServerError,
+                (
+                    ContentType::JSON,
+                    make_bad_json_data_response(format!("Could not write to {}: {}", ipath, e)),
+                ),
+            )
+        };
+        status::Custom(
+            Status::Ok,
+            (
+                ContentType::JSON,
+                make_good_json_data_response("ok".to_string()),
+            ),
+        )
+    } else {
+        status::Custom(
+            Status::BadRequest,
+            (
+                ContentType::JSON,
+                make_bad_json_data_response("bad repo path".to_string()),
+            ),
+        )
+    }
+}
