+++ src/endpoints/mod.rs
@@ -5,9 +5,9 @@ pub mod i18n;
 pub mod sse;
 pub mod navigation;
 pub mod gitea;
-pub mod burrito;
 pub mod clients;
 pub(crate) mod error;
 pub mod app_state;
 pub mod content_utils;
 pub mod git2;
+pub mod burrito2;
